Moving scheme work into importer
