+++ src/main/resources/themes/one_dark.xml
@@ -3,6 +3,7 @@
     <property name="oneDarkScheme">bundled</property>
   </metaInfo>
 <!--  Please do not edit this file it will be overwritten by the construction process -->
+<!--  This just prevents the flash of color when the IDE first loads :)-->
   <colors>
     <option name="ADDED_LINES_COLOR" value="98c379"/>
     <option name="ANNOTATIONS_COLOR" value="a0a7b4"/>
